requirements.txt
aiohttp==3.7.2
alpha-vantage==2.2.0
argon2-cffi==20.1.0
asgiref 
async-generator==1.10
async-timeout==3.0.1
attrs==20.2.0
backcall==0.2.0
bleach==3.2.1
certifi==2020.12.5
decorator==4.4.2
defusedxml==0.6.0
Django==3.0.5
django-crispy-forms==1.10.0
djongo==1.3.3
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
mistune==0.8.4
multidict==5.0.0
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.2
notebook==6.1.4
packaging==20.4
pandocfilters==1.4.3
parso==0.7.1
pickleshare==0.7.5
prometheus-client==0.8.0
prompt-toolkit==3.0.8
# psycopg2
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.7.2
pymongo==3.11.2
pyparsing==2.4.7
pyrsistent==0.17.3
pytz
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
Send2Trash==1.5.0
sqlparse==0.2.4
terminado==0.9.1
testpath==0.4.4
tornado==6.0.4
traitlets==5.0.5
typing-extensions==3.7.4.3
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
yarl==1.6.2
